Fixed a bug that affected changing the ancestor state during search
In certain situations, the interactions between the temporal framework
and the change_ancesstor function could lead to have the same
"start-action" added repeated to the running_actions list. This caused
that, while extracting the time when the "end-action" should be applied,
the one from the removed ancesstor could be chosen, leading to not sound
temporal plans.
This could not happen before the timed goals framework implementation.